# Switch to MessagePack