--- a/HPXMLtoOpenStudio/resources/meta_measure.rb
+++ b/HPXMLtoOpenStudio/resources/meta_measure.rb
@@ -134,10 +134,12 @@ def run_hpxml_workflow(rundir, measures, measures_dir, debug: false, output_vars
 
   # Apply reporting measures
   runner.setLastEnergyPlusSqlFilePath(File.join(rundir, 'eplusout.sql'))
+  runner.setLastEpwFilePath(File.join(rundir, 'in.epw'))
   success = apply_measures(measures_dir, measures, runner, model, false, 'OpenStudio::Measure::ReportingMeasure')
   report_measure_errors_warnings(runner, rundir, debug)
   report_os_warnings(os_log, rundir)
   runner.resetLastEnergyPlusSqlFilePath
+  runner.resetLastEpwFilePath
 
   Dir[File.join(rundir, 'results_*.csv')].each do |csv_path|
     print "#{print_prefix}Wrote output file: #{csv_path}.\n"
@@ -364,6 +366,7 @@ def run_measure(model, measure, argument_map, runner)
     if measure.class.superclass.name.to_s == 'OpenStudio::Measure::ReportingMeasure'
       runner_child.setLastOpenStudioModel(model)
       runner_child.setLastEnergyPlusSqlFilePath(runner.lastEnergyPlusSqlFile.get.path)
+      runner_child.setLastEpwFilePath(runner.lastEpwFilePath.get)
       measure.run(runner_child, argument_map)
     else
       measure.run(model, runner_child, argument_map)