--- a/HPXMLtoOpenStudio/resources/output.rb
+++ b/HPXMLtoOpenStudio/resources/output.rb
@@ -139,7 +139,9 @@ class WT
 end
 
 class OutputMethods
-  def self.get_timestamps(timeseries_frequency, sqlFile, hpxml, timestamps_local_time = nil)
+  def self.get_timestamps(timeseries_frequency, msgpackData, hpxml, timestamps_local_time = nil)
+    return if msgpackData.nil?
+
     if timeseries_frequency == 'hourly'
       interval_type = 1
     elsif timeseries_frequency == 'daily'
@@ -150,9 +152,13 @@ class OutputMethods
       interval_type = -1
     end
 
-    query = "SELECT Year || ' ' || Month || ' ' || Day || ' ' || Hour || ' ' || Minute As Timestamp FROM Time WHERE IntervalType='#{interval_type}'"
-    values = sqlFile.execAndReturnVectorOfString(query)
-    fail "Query error: #{query}" unless values.is_initialized
+    if msgpackData.keys.include? 'MeterData'
+      # Get data for ReportUtilityBills measure
+      ep_timestamps = msgpackData['MeterData']['Monthly']['Rows'].map { |r| r.keys[0] }
+    else
+      # Get data for other reporting measures
+      ep_timestamps = msgpackData['Rows'].map { |r| r.keys[0] }
+    end
 
     if timestamps_local_time == 'DST'
       dst_start_ts = Time.utc(hpxml.header.sim_calendar_year, hpxml.header.dst_begin_month, hpxml.header.dst_begin_day, 2)
@@ -163,8 +169,11 @@ class OutputMethods
     end
 
     timestamps = []
-    values.get.each do |value|
-      year, month, day, hour, minute = value.split(' ')
+    year = hpxml.header.sim_calendar_year.to_s # Not available in output timestamps
+    ep_timestamps.each do |ep_timestamp|
+      month_day, hour_minute = ep_timestamp.split(' ')
+      month, day = month_day.split('/')
+      hour, minute, _ = hour_minute.split(':')
       ts = Time.utc(year, month, day, hour, minute)
 
       if timestamps_local_time == 'DST'
@@ -183,11 +192,12 @@ class OutputMethods
     return timestamps
   end
 
-  def self.teardown(sqlFile)
-    sqlFile.close()
-
-    # Ensure sql file is immediately freed; otherwise we can get
-    # errors on Windows when trying to delete this file.
-    GC.start()
+  def self.msgpack_frequency_map
+    return {
+      'timestep' => 'TimeStep',
+      'hourly' => 'Hourly',
+      'daily' => 'Daily',
+      'monthly' => 'Monthly',
+    }
   end
 end