--- a/HPXMLtoOpenStudio/resources/output.rb
+++ b/HPXMLtoOpenStudio/resources/output.rb
@@ -139,7 +139,7 @@ class WT
 end
 
 class OutputMethods
-  def self.get_timestamps(timeseries_frequency, msgpackData, hpxml, timestamps_local_time = nil)
+  def self.get_timestamps(timeseries_frequency, msgpackData, hpxml, add_dst_column = false, add_utc_column = false)
     return if msgpackData.nil?
 
     if timeseries_frequency == 'hourly'
@@ -160,36 +160,43 @@ class OutputMethods
       ep_timestamps = msgpackData['Rows'].map { |r| r.keys[0] }
     end
 
-    if timestamps_local_time == 'DST'
+    if add_dst_column
       dst_start_ts = Time.utc(hpxml.header.sim_calendar_year, hpxml.header.dst_begin_month, hpxml.header.dst_begin_day, 2)
       dst_end_ts = Time.utc(hpxml.header.sim_calendar_year, hpxml.header.dst_end_month, hpxml.header.dst_end_day, 1)
-    elsif timestamps_local_time == 'UTC'
+    end
+    if add_utc_column
       utc_offset = hpxml.header.time_zone_utc_offset
       utc_offset *= 3600 # seconds
     end
 
     timestamps = []
-    year = hpxml.header.sim_calendar_year.to_s # Not available in output timestamps
+    timestamps_dst = [] if add_dst_column
+    timestamps_utc = [] if add_utc_column
+    year = hpxml.header.sim_calendar_year.to_s
     ep_timestamps.each do |ep_timestamp|
       month_day, hour_minute = ep_timestamp.split(' ')
       month, day = month_day.split('/')
       hour, minute, _ = hour_minute.split(':')
       ts = Time.utc(year, month, day, hour, minute)
 
-      if timestamps_local_time == 'DST'
+      timestamps << ts.iso8601.delete('Z')
+
+      if add_dst_column
         if (ts >= dst_start_ts) && (ts < dst_end_ts)
-          ts += 3600 # 1 hr shift forward
+          ts_dst = ts + 3600 # 1 hr shift forward
+        else
+          ts_dst = ts
         end
-      elsif timestamps_local_time == 'UTC'
-        ts -= utc_offset
+        timestamps_dst << ts_dst.iso8601.delete('Z')
       end
 
-      ts_iso8601 = ts.iso8601
-      ts_iso8601 = ts_iso8601.delete('Z') if timestamps_local_time != 'UTC'
-      timestamps << ts_iso8601
+      if add_utc_column
+        ts_utc = ts - utc_offset
+        timestamps_utc << ts_utc.iso8601
+      end
     end
 
-    return timestamps
+    return timestamps, timestamps_dst, timestamps_utc
   end
 
   def self.msgpack_frequency_map