--- a/ReportHPXMLOutput/measure.rb
+++ b/ReportHPXMLOutput/measure.rb
@@ -31,6 +31,7 @@ class ReportHPXMLOutput < OpenStudio::Measure::ReportingMeasure
     format_chs = OpenStudio::StringVector.new
     format_chs << 'csv'
     format_chs << 'json'
+    format_chs << 'msgpack'
     arg = OpenStudio::Measure::OSArgument::makeChoiceArgument('output_format', format_chs, false)
     arg.setDisplayName('Output Format')
     arg.setDescription('The file format of the annual (and timeseries, if requested) outputs.')
@@ -160,9 +161,9 @@ class ReportHPXMLOutput < OpenStudio::Measure::ReportingMeasure
       results_out << ["#{key} (#{cost_mult.units})", cost_mult.output.round(2)]
     end
 
-    if output_format == 'csv'
+    if ['csv'].include? output_format
       CSV.open(output_path, 'wb') { |csv| results_out.to_a.each { |elem| csv << elem } }
-    elsif output_format == 'json'
+    elsif ['json', 'msgpack'].include? output_format
       h = {}
       results_out.each do |out|
         next if out == [line_break]
@@ -172,8 +173,12 @@ class ReportHPXMLOutput < OpenStudio::Measure::ReportingMeasure
         h[grp][name.strip] = out[1]
       end
 
-      require 'json'
-      File.open(output_path, 'w') { |json| json.write(JSON.pretty_generate(h)) }
+      if output_format == 'json'
+        require 'json'
+        File.open(output_path, 'w') { |json| json.write(JSON.pretty_generate(h)) }
+      elsif output_format == 'msgpack'
+        File.open(output_path, 'w') { |json| h.to_msgpack(json) }
+      end
     end
     runner.registerInfo("Wrote hpxml output to #{output_path}.")
   end