--- a/ReportHPXMLOutput/measure.rb
+++ b/ReportHPXMLOutput/measure.rb
@@ -64,7 +64,7 @@ class ReportHPXMLOutput < OpenStudio::Measure::ReportingMeasure
       runner.registerError('Cannot find eplusout.msgpack.')
       return false
     end
-    @msgpackData = MessagePack.unpack(File.read(File.join(output_dir, 'eplusout.msgpack')))
+    @msgpackData = MessagePack.unpack(File.read(File.join(output_dir, 'eplusout.msgpack'), mode: 'rb'))
 
     hpxml_defaults_path = model.getBuilding.additionalProperties.getFeatureAsString('hpxml_defaults_path').get
     hpxml = HPXML.new(hpxml_path: hpxml_defaults_path)