--- a/ReportSimulationOutput/measure.rb
+++ b/ReportSimulationOutput/measure.rb
@@ -433,6 +433,13 @@ class ReportSimulationOutput < OpenStudio::Measure::ReportingMeasure
 
     output_dir = File.dirname(runner.lastEpwFilePath.get.to_s)
 
+    hpxml_path = @model.getBuilding.additionalProperties.getFeatureAsString('hpxml_path').get
+    hpxml_defaults_path = @model.getBuilding.additionalProperties.getFeatureAsString('hpxml_defaults_path').get
+    building_id = @model.getBuilding.additionalProperties.getFeatureAsString('building_id').get
+    @hpxml = HPXML.new(hpxml_path: hpxml_defaults_path, building_id: building_id)
+    HVAC.apply_shared_systems(@hpxml) # Needed for ERI shared HVAC systems
+    @eri_design = @hpxml.header.eri_design
+
     setup_outputs()
 
     @msgpackData = MessagePack.unpack(File.read(File.join(output_dir, 'eplusout.msgpack')))
@@ -444,13 +451,6 @@ class ReportSimulationOutput < OpenStudio::Measure::ReportingMeasure
     if not @emissions.empty?
       @msgpackDataHourly = MessagePack.unpack(File.read(File.join(output_dir, 'eplusout_hourly.msgpack')))
     end
-
-    hpxml_path = @model.getBuilding.additionalProperties.getFeatureAsString('hpxml_path').get
-    hpxml_defaults_path = @model.getBuilding.additionalProperties.getFeatureAsString('hpxml_defaults_path').get
-    building_id = @model.getBuilding.additionalProperties.getFeatureAsString('building_id').get
-    @hpxml = HPXML.new(hpxml_path: hpxml_defaults_path, building_id: building_id)
-    HVAC.apply_shared_systems(@hpxml) # Needed for ERI shared HVAC systems
-    @eri_design = @hpxml.header.eri_design
 
     # Set paths
     if not @eri_design.nil?