## Pull Request Description

Closes #903. Replaces use of EnergyPlus SQL output with MessagePack for performance & file size reasons.

Here are results for various simulations. Runtime refers to the entire end-to-end workflow (creation of IDF, simulation, and output processing). Filesize refers to just the sql/msgpack output files.

  | Prev Runtime [s] | New Runtime [s] | Runtime Reduction | Prev Filesize [MB] | New Filesize [MB] | Filesize Reduction
-- | -- | -- | -- | -- | -- | --
Simulation w/ 60-min timestep, no timeseries output | 5.0 | 5.0 | 0% | 0.5 | 0.1 | 80%
Simulation w/ 60-min timestep, 60-min outputs (end uses only) | 6.0 | 5.5 | 8% | 8.2 | 1.6 | 80%
Simulation w/ 60-min timestep, 60-min outputs (ALL outputs) | 10.7 | 7.5 | 30% | 28.1 | 5.0 | 82%
Simulation w/ 10-min timestep, no timeseries output | 15.2 | 15.1 | 1% | 0.5 | 0.1 | 80%
Simulation w/ 10-min timestep, 10-min outputs (end uses only) | 21.1 | 17.9 | 15% | 47.5 | 8.4 | 82%
Simulation w/ 10-min timestep, 10-min outputs (ALL outputs) | 47.6 | 29.8 | 37% | 171.7 | 29.2 | 83%

This PR also now allows OS-HPXML output files (e.g., results_timeseries.*) to be in MessagePack format (instead of CSV/JSON) if chosen.

## Checklist

PR Author: Check these when they're done. Not all may apply. ~~strikethrough~~ and check any that do not apply. 

PR Reviewer: Verify each has been completed.

- [x] ~Schematron validator (`EPvalidator.xml`) has been updated~
- [x] ~Sample files have been added/updated (via `tasks.rb`)~
- [x] ~Unit tests have been added/updated (e.g., `HPXMLtoOpenStudio/tests`)~
- [x] Documentation has been updated
- [x] Changelog has been updated
- [x] `openstudio tasks.rb update_measures` has been run
- [x] ~Checked the code coverage report on CI~
- [x] No unexpected changes to simulation results of sample files