@shorowit 2022-04-27
This provides a nice speed improvement but we're unable to proceed because [it's not currently working on Windows](https://github.com/msgpack/msgpack-ruby/issues/278).